Move packages of tests
Had a bit of a refactor of the service package a few weeks ago but looks
like the tests did not get move to the coresponding test packages. This
should correct this.
