+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/service/AmendPanelMembersServiceTest.java
@@ -1,9 +1,7 @@
 package uk.gov.hmcts.reform.sscscorbackend.service;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.*;
 
 import org.junit.Before;
 import org.junit.Test;
