+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/service/QuestionsEmailMessageBuilderTest.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.sscscorbackend.service;
 
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertThat;
 
 import org.junit.Test;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
