+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/service/pdf/DecodeJsonUtilTest.java
@@ -1,11 +1,10 @@
-package uk.gov.hmcts.reform.sscscorbackend.service;
+package uk.gov.hmcts.reform.sscscorbackend.service.pdf;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.junit.Assert.assertThat;
 
 import org.junit.Test;
-import uk.gov.hmcts.reform.sscscorbackend.service.pdf.DecodeJsonUtil;
 
 public class DecodeJsonUtilTest {
     @Test
