+++ src/test/java/uk/gov/hmcts/reform/sscscorbackend/service/pdf/I18nBuilderTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscscorbackend.service;
+package uk.gov.hmcts.reform.sscscorbackend.service.pdf;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -6,7 +6,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import java.io.IOException;
 import java.util.Map;
 import org.junit.Test;
-import uk.gov.hmcts.reform.sscscorbackend.service.pdf.I18nBuilder;
 
 public class I18nBuilderTest {
     @Test
